+++ src/main/java/com/backend/notice/presentation/ActivityNoticeController.java
@@ -1,0 +1,27 @@
+package com.backend.notice.presentation;
+
+import com.backend.common.response.CommonResponse;
+import com.backend.notice.application.NoticeStrategyFactory;
+import com.backend.notice.dto.NoticeResponse;
+import com.backend.notice.presentation.status.NoticeCategory;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/activity-notice")
+@RequiredArgsConstructor
+public class ActivityNoticeController {
+
+    private final NoticeStrategyFactory noticeStrategyFactory;
+
+    @GetMapping("/all")
+    public CommonResponse<List<NoticeResponse>> getActivityNotice() {
+        return CommonResponse.success(noticeStrategyFactory.getAllNotices(NoticeCategory.from("activity")), HttpStatus.OK, "대외활동/아르바이트 공고 조회 성공");
+    }
+
+}
